requirements.txt
absl-py==1.4.0
albumentations==1.3.1
asttokens==2.2.1
backcall==0.2.0
batchgenerators==0.24
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dicom2nifti==2.4.7
einops==0.7.0
et-xmlfile==1.1.0
executing==1.2.0
flit_core==3.6.0
fonttools==4.39.2
future==0.18.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.51.3
hiddenlayer==0.3
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.0
nibabel==5.0.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
prompt-toolkit==3.0.38
protobuf==4.22.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-gdcm==3.0.21
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
setuptools==65.6.3
SimpleITK==2.2.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tifffile==2023.2.28
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.9.0
typing_extensions==4.4.0
unittest2==1.1.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
xlwt==1.3.0
zipp==3.15.0
